pro1: devicesettings: Disable custom keymap pref if no file exists
This setting was implemented with advanced users in mind, implying they
would create a custom keymap (under /data/system/keyboard/keymap).
However, regular users wouldn't really know what the setting was for
and toggling the switch was doint nothing.

+++ devicesettings/src/org/lineageos/settings/device/KeyboardSettingsFragment.java
@@ -104,11 +104,23 @@ public class KeyboardSettingsFragment extends PreferenceFragment
         SwitchPreference remapFnSpacePowerPref =
                 findPreference(Constants.KEYBOARD_KEYMAP_SPACEPOWER_KEY);
 
-        if (customKeymapPref.isChecked()) {
-            String text = readFile(Constants.KEYBOARD_KEYMAP_CFG_FILE);
-            writeFile(Constants.KEYBOARD_KEYMAP_SYS_FILE, text);
-        }
-        else {
+        File customKeymapFile = new File(Constants.KEYBOARD_KEYMAP_CFG_FILE);
+        if (customKeymapFile.exists()) {
+            customKeymapPref.setEnabled(true);
+
+            if (customKeymapPref.isChecked()) {
+                remapFnKeysPref.setEnabled(false);
+                remapFnSpacePowerPref.setEnabled(false);
+
+                String text = readFile(Constants.KEYBOARD_KEYMAP_CFG_FILE);
+                writeFile(Constants.KEYBOARD_KEYMAP_SYS_FILE, text);
+            }
+        } else {
+            customKeymapPref.setEnabled(false);
+            customKeymapPref.setChecked(false);
+            remapFnKeysPref.setEnabled(true);
+            remapFnSpacePowerPref.setEnabled(true);
+
             int i;
             if (remapFnSpacePowerPref.isChecked()) {
                 for (i = 0; i < Constants.KEYBOARD_KEYMAP_SPACEPOWER_TEXT.length; ++i) {
